interviewer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.8
elevenlabs==0.2.26
python-dotenv==1.0.0
python-multipart==0.0.6
